swing

+++ Homework1 - copia/CirculoCopia.java
@@ -3,7 +3,7 @@ import java.lang.Math;
 
   public class CirculoCopia{
     public static void main(String[] args) {
-      double R =Integer.parseInt(JOptionPane.showInputDialog(null, "Circunferencia de un circulo" + "\n" +
+      double R =Double.parseDouble(JOptionPane.showInputDialog(null, "Circunferencia de un circulo" + "\n" +
       "\t" + "Ingresa el radio "));
       double Area = Math.PI*(Math.pow (R, 2));
       JOptionPane.showMessageDialog(null, "El area es: " + Area );
