+++ Homework2 - copia/TelasCopia.java
@@ -2,7 +2,7 @@ import javax.swing.JOptionPane;
 
   public class TelasCopia{
     public static void main(String[] args) {
-      double CM = Integer.parseInt(JOptionPane.showInputDialog("Conversion de metros a pulgadas por rollo de tela" + "\n" +
+      double CM = Double.parseDouble(JOptionPane.showInputDialog("Conversion de metros a pulgadas por rollo de tela" + "\n" +
       "\t" + "Ingresa la cantidad de metros que requiera"));
       double PG = CM / 0.0254;
       JOptionPane.showMessageDialog(null, "Se necesitaran: " + PG + " pulgadas");
